реструктуризация и переписывание тестов
